Переделал fetchAndRnd и прочее
