+++ index.html
@@ -14,6 +14,5 @@
 </div>
 <a href=""></a>
 </body>
-<script type="module" src="./render.js"></script>
 <script type="module" src="./main.js"></script>
 </html>
