+++ render.js
@@ -1,57 +1,10 @@
-import { initLikeListener,quoteCommets, commentList, fetchAndRenderComments, addComment } from "./main.js";
+import {commentList} from "./main.js";
 import { initDeleteButtonsListeners } from "./delbutton.js";
-import { setToken, token, loginPost, getComments, setUser} from "./api.js";
+import { token, postComment } from "./api.js";
+import { renderLoginForm } from "./renderLogin.js";
+import { sanitizeHtml } from './sanitizeHtml.js';
 
-// Сохраняем данные об объекте User в localStorage
-// Сохраняем данные об объекте User в localStorage
 
-
-//Рендерим форму входа
-export const renderLoginForm = () => {
-  const appHtml = document.getElementById("app");
-  const loginHtml = `
-  <div class="add-form">
-    <input 
-    type="text"
-    id="login-input" 
-    class="add-form-name"
-    placeholder="Логин"
-    />
-    <input 
-    type="text"
-    id="password-input"
-    class="add-form-name"
-    placeholder="Пароль"
-    />
-    <button id="login-form-button" class="add-form-button">Войти</button>
-  </div>`;
-  appHtml.innerHTML = loginHtml;
-
-  //Добавляем действие по клику на "авторизация"
-  const buttonLoginElement = document.getElementById("login-form-button");
-  const loginInputElement = document.getElementById("login-input");
-  const passwordInputElement = document.getElementById("password-input");
-
-  buttonLoginElement.addEventListener("click", (event) => {
-    event.preventDefault();
-    if (!loginInputElement.value || !passwordInputElement.value) {
-      alert("Проверьте оба поля  на заполненность");
-      return
-    }
-    loginPost({
-      login: loginInputElement.value,
-      password: passwordInputElement.value,
-    }).then((responseData) => {
-      localStorage.setItem("token", responseData.user.token);
-      localStorage.setItem("user", JSON.stringify(responseData.user));
-      setToken(responseData.user.token);
-      setUser(responseData.user);
-      console.log(token);
-    }).then(() => {
-      renderComments();
-    })
-  });
-};
 
 //Выводим комменты
 export const renderComments = () => {
@@ -80,12 +33,12 @@ export const renderComments = () => {
 
   //Форма ввода комментария
   const contentHtml = () => {
-
     const btnLogin = `
     <p class="render-login-btn">  Чтобы добавить комментарий, 
     <a id="render-login-btn">авторизуйтесь</a> </p>`
 
-    if (!token) return `<ul id="comments" class="comments">${commentsHtml}</ul> ${btnLogin}`;
+    if (!token) return `<ul id="comments" class="comments">${commentsHtml}</ul>
+     ${btnLogin}`;
     return `<ul id="comments" class="comments">${commentsHtml}</ul>
     <div id="add-form" class="add-form">
       <input id="add-name" type="text" class="add-form-name" placeholder="Введите ваше имя" />
@@ -97,17 +50,113 @@ export const renderComments = () => {
     </div>`
   }
 
-  initLikeListener();
+  /* initLikeListener();
   initDeleteButtonsListeners();
-  quoteCommets();
+  quoteCommets(); */
   appHtml.innerHTML = contentHtml()
 
   //Переход к форме авторизации по клику
-  const buttonElement = document.getElementById("render-login-btn");
-  buttonElement.addEventListener("click", (event) => {
-    event.preventDefault();
-    renderLoginForm();
-  });
+  setLoginBtn = () => {
+    const buttonLoginElement = document.getElementById("render-login-btn");
+    buttonLoginElement.addEventListener("click", (event) => {
+      event.preventDefault();
+      renderLoginForm();
+    });
+  };
+
+  setLoginBtn();
 
 
 };
+//Активность кнопки лайк
+export const initLikeListener = () => {
+  const buttonLike = document.querySelectorAll(".like-button");
+  for (const iteratorLike of buttonLike) {
+    iteratorLike.addEventListener("click", (event) => {
+      event.stopPropagation();
+      const index = iteratorLike.dataset.index;
+      commentList[index].likes += commentList[index].isLiked ? -1 : +1;
+      commentList[index].isLiked = !commentList[index].isLiked;
+      renderComments(); //перерисовываем форму для лайков с счетчиком
+    });
+  }
+};
+
+//Цитирование
+const quoteCommets = () => {
+  const textAreaElement = document.getElementById("add-text");
+  const commentElements = document.querySelectorAll(".comment");
+  for (const commentElement of commentElements) {
+    commentElement.addEventListener("click", () => {
+      const index = commentElement.dataset.index;
+      const commentText = commentList[index].text;
+      const commentAuthor = commentList[index].name;
+      textAreaElement.value = `${commentText} > ${commentAuthor}`;
+    })
+  };
+
+};
+
+/* addComment(); */
+
+
+ addComment = () => {
+  const textAreaElement = document.getElementById("add-text");
+  const inputElement = document.getElementById("add-name");
+  const buttonElement = document.getElementById("add-form-button");
+  console.log(inputElement, textAreaElement)
+  buttonElement.addEventListener("click", () => {
+    inputElement.classList.remove("error");
+    if (inputElement.value === "") {
+      inputElement.classList.add("error");
+    }
+    if (textAreaElement.value === "") {
+      textAreaElement.classList.add("error");
+      return;
+    };
+
+    //2.13. надпись о загрузке коммента и блокировка кнопки "добавить".
+    postComment(inputElement.value,
+      textAreaElement.value, sanitizeHtml(textareaInputElement.value))
+      .then((response) => {
+        if (response.status === 201) {
+          return response.json();
+        }
+        if (response.status === 400) {
+          throw new Error("Некорректный запрос error 400");
+        } if (response.status === 500) {
+          throw new Error("Ошибка сервера error 500");
+        }
+      }).then(() => {
+        inputElement.value = "";
+        textAreaElement.value = "";
+        return getComments();
+      })
+      .catch((error) => {
+        buttonElement.disabled = true;
+        if (error.message === "Некорректный запрос error 400") {
+          alert("Длина имени не может быть меньше 3 символов");
+        } else if (error.message === "Ошибка сервера error 500") {
+          alert("Ошибка сервера");
+        } else if (error.message === "Failed to fetch") {
+          alert("Отуствует соединение к интеренету");
+        };
+        buttonElement.disabled = false;
+        renderComments();
+
+      })
+  })
+  if (token) {
+    const buttonElement = document.getElementById('add-button');
+    buttonElement.addEventListener('click', addNewComment);
+  }
+  initLikeListener();
+  initDeleteButtonsListeners
+  quoteCommets();
+};
+
+
+
+
+
+
