+++ sanitizeHtml.js
@@ -1,0 +1,3 @@
+export const sanitizeHtml = (htmlString) => {
+    return htmlString.replaceAll("<", "&lt;").replaceAll(">", "&gt;");
+};
